Data.DotNet.TimeSpan.Internal

Description

.NET TimeSpan implemented in Haskell.

Documentation

newtype TimeSpan

.NET TimeSpan: Represents a time interval.

Constructors

TimeSpan Int64

millisPerSecond :: Int64

millisPerMinute :: Int64

millisPerHour :: Int64

millisPerDay :: Int64

ticksPerMillisecond :: Int64

ticksPerSecond :: Int64

ticksPerMinute :: Int64

ticksPerHour :: Int64

ticksPerDay :: Int64

daysPerTick :: Double

hoursPerTick :: Double

minutesPerTick :: Double

secondsPerTick :: Double

millisPerTick :: Double

maxMillis :: Int64

minMillis :: Int64

timeSpanTicks :: Int64 -> TimeSpan

Initializes a new instance of the TimeSpan structure to the specified number of ticks.

timeSpanHoursMinsSecs :: Int64 -> Int64 -> Int64 -> TimeSpan

Initializes a new instance of the TimeSpan structure to a specified number of hours, minutes, and seconds.

timeSpanDaysHoursMinsSecs :: Int64 -> Int64 -> Int64 -> Int64 -> TimeSpan

Initializes a new instance of the TimeSpan structure to a specified number of days, hours, minutes, and seconds.

timeSpanDaysHoursMinsSecsMillis :: Int64 -> Int64 -> Int64 -> Int64 -> Int64 -> TimeSpan

Initializes a new instance of the TimeSpan structure to a specified number of days, hours, minutes, seconds, and milliseconds.

ticks :: TimeSpan -> Int64

Gets the number of ticks that represent the value of the current TimeSpan structure.

days :: TimeSpan -> Int64

Gets the days component of the time interval represented by the current TimeSpan structure.

hours :: TimeSpan -> Int64

Gets the hours component of the time interval represented by the current TimeSpan structure.

minutes :: TimeSpan -> Int64

Gets the minutes component of the time interval represented by the current TimeSpan structure.

seconds :: TimeSpan -> Int64

Gets the seconds component of the time interval represented by the current TimeSpan structure.

millis :: TimeSpan -> Int64

Gets the milliseconds component of the time interval represented by the current TimeSpan structure.

fromSeconds :: Double -> TimeSpan

Returns a TimeSpan that represents a specified number of seconds, where the specification is accurate to the nearest millisecond.

fromMinutes :: Double -> TimeSpan

Returns a TimeSpan that represents a specified number of minutes, where the specification is accurate to the nearest millisecond.

fromHours :: Double -> TimeSpan

Returns a TimeSpan that represents a specified number of hours, where the specification is accurate to the nearest millisecond.

fromDays :: Double -> TimeSpan

Returns a TimeSpan that represents a specified number of days, where the specification is accurate to the nearest millisecond.

totalDays :: TimeSpan -> Double

Gets the value of the current TimeSpan structure expressed in whole and fractional days.

totalHours :: TimeSpan -> Double

Gets the value of the current TimeSpan structure expressed in whole and fractional hours.

totalMinutes :: TimeSpan -> Double

Gets the value of the current TimeSpan structure expressed in whole and fractional minutes.

totalSeconds :: TimeSpan -> Double

Gets the value of the current TimeSpan structure expressed in whole and fractional seconds.

totalMillis :: TimeSpan -> Double

Gets the value of the current TimeSpan structure expressed in whole and fractional milliseconds.

data FormatLiteral

Constructors

Positive

Negative

padded :: Int -> a -> [a] -> [a]

timeSpanString :: TimeSpan -> String

interval :: Double -> Int64 -> TimeSpan